merge new v5

+++ src/main/java/com/lemon/supershop/swp391fa25evdm/category/model/dto/CategoryReq.java
@@ -5,7 +5,7 @@ public class CategoryReq {
     private String brand;
     private String version;
     private String type;
-    private Double basePrice;
+    private Long basePrice;
     private Integer warranty;
     private Boolean isSpecial;
     private String description;
@@ -46,11 +46,11 @@ public class CategoryReq {
         this.type = type;
     }
 
-    public Double getBasePrice() {
+    public Long getBasePrice() {
         return basePrice;
     }
 
-    public void setBasePrice(Double basePrice) {
+    public void setBasePrice(Long basePrice) {
         this.basePrice = basePrice;
     }
